+++ server-side/src/main/java/com/tick42/quicksilver/repositories/base/FileRepository.java
@@ -3,6 +3,6 @@ package com.tick42.quicksilver.repositories.base;
 import com.tick42.quicksilver.models.File;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FileRepository extends JpaRepository<File, Integer> {
+public interface FileRepository extends JpaRepository<File, Long> {
     File findByName(String name);
 }
